refactor: Presentable 프로토콜 분리
