+++ NewsFit/NewsFit/View/Protocols/NewsDescriptionPresentable.swift
@@ -1,0 +1,3 @@
+protocol NewsDescriptionPresentable {
+  var body: String { get }
+}
