+++ NewsFit/NewsFit/View/Protocols/PressImagePresentable.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol PressImagePresentable {
+  var pressImageURL: URL? { get }
+}
